feat(scalar-range): widen --from/--to to u128 for address discovery
User's literal input
  --address 1PWo3JeB9jrGwfHDNpdGK54CRas7fsVzXU \
  --from 0x400000000000000000 --to 0x7fffffffffffffffff
was rejected with `hex "...number too large to fit in target type"`
because the scalar range was u64-bounded (2^64 - 1 ≈ 1.84e19) while the
user's `0x400000000000000000` (2^70 ≈ 1.18e21) exceeds it.

Widen the scalar range for address-discovery to u128:

Config:
  range_from:  Option<u64> -> Option<u128>
  range_to:    Option<u64> -> Option<u128>
  MAX_SEARCH:  u64           -> u128         (u128::MAX)
  try_with_range: u64 × u64    -> u128 × u128

Search:
  SearchMatch.j:  u64  -> u128  (release currently [Unreleased];
                                wider scalar recovery is necessary
                                for the address-mode path)
  VariantIndex::match_x(j: u64)        -> u128
  search::sweep_address(start, end, ...)
       start: u64 -> u128
       end:   u64 -> u128
  sweep_cached(index, cache_path, start_j: u64) -> u128

The +G chain in sweep_address walks the range in u128; each step
the math is `chunk_start + c` where `chunk_start: u128` and
`c: u128`, advancing by `+= ecc::generator()` (a constant). The
geometric iteration matches the variant-mode sweep's shape but
doesn't need the variant X-coordinate or binary-search machinery
that the variant-mode sweep uses (it terminates with a hash40
compare per step).

Ponytail-keeping-what-works: variant-mode (sweep_parallel,
sweep_and_cache, sweep_cached, BinaryCacheWriter) stays in u64.
Only sweep_address widens. The variant-mode has a fundamentally
u64-bounded design (cache file names, checkpoints, binary cache
format all encode u64 chunk indices) and lifting it would be a
much bigger structural change than this slice.

Tests adjusted:
  - tests/orchestrator.rs: `assert_eq!(match_.j, d)` -> `match_.j,
    u128::from(d)` (d is still u64 since d < u64::MAX in tests).
  - examples/sweep_bench.rs: the bench sink used `wrapping_add(m.j)`
    which now needs u64 conversion; saturated via `try_from(...).unwrap_or(u64::MAX)`.

Verified: cargo fmt --check clean; cargo clippy --all-targets
--all-features -- -D warnings clean; cargo test --lib 79/79 pass;
cargo test --bin find 4/4 pass; integration + KAT + differential +
audit + orchestrator suites 47/47 pass; example sweep_bench
compiles and benchmarks correctly.

Empirical demonstration that the user's input now works:
  $ ./target/release/find --address 1PWo3JeB9jrGwfHDNpdGK54CRas7fsVzXU \
        --from 0x400000000000000000 --to 0x7fffffffffffffffff
  [INFO  address mode: target=f6f5... range=[1180591620717411303424,
         2361183241434822606847]]

The range parses. The sweep runs. The expected work is
range_width = ~1.18e21 iterations × ~1 µs/scalar ≈ 1.18e15 core-seconds
on a single M3 Pro (~37,000 years). The address-mode pipeline does
now process the user's input — the search is just infeasibly
expensive at typical hardware budgets, as documented in ADR-0011.

In contrast, the small-range smoke test (the same address-mode
pipeline against `d=5`) finds the match in ~1.5 ms:
  $ ./target/release/find --address 17Vu7st1U1KwymUKU4jJheHHGRVNqrcfLD \
        --from 1 --to 10
  MATCH DISCOVERED (Variant: address/d)
  Search scalar j: 5
  Total Search Duration: 1.510958ms

+++ src/config.rs
@@ -41,8 +41,16 @@ pub const TRILLION: u64 = 1_000_000_000_000;
 /// Each chunk corresponds to ~32 GB of binary cache on disk.
 pub const DEFAULT_CACHE_CHUNK_SIZE: u64 = 1_000_000_000;
 
-/// Theoretical maximum search boundary for 64-bit scalars.
-pub const MAX_SEARCH: u64 = u64::MAX;
+/// Theoretical maximum search boundary for scalars.
+///
+/// `u128::MAX` covers the entire Secp256k1 scalar field
+/// (`2^256 mod n < 2^128` doesn't quite hold — `n < 2^256`, so the largest
+/// canonical scalar in the field is `n - 1 < 2^256`, which is bigger than
+/// `u128::MAX = 2^128 - 1`). The address-discovery slice uses `u128` so
+/// the user's `0x4000…0000` style inputs are accepted and the chain
+/// loop can advance by `+ G` past `u64::MAX`. Very-near-`2^128` ranges
+/// are still rejected at parse time via `u128::from_str_radix`.
+pub const MAX_SEARCH: u128 = u128::MAX;
 
 /// Minimum non-zero search scalar.
 ///
@@ -195,13 +203,19 @@ pub struct Config {
     /// `MIN_SEARCH_SCALAR`". Combined with `range_to`, scopes the entire
     /// session to a single user-specified window without persisted
     /// checkpoints (see ADR-0011).
-    pub range_from: Option<u64>,
+    ///
+    /// **Type: `u128`.** Values above `2^64 - 1` are accepted so the
+    /// `--address` path can handle very-wide user ranges such as the
+    /// form-template input `0x400000000000000000:0x7fffffffffffffffff`.
+    pub range_from: Option<u128>,
     /// Optional inclusive-upper scalar bound for the sweep.
     ///
     /// `Some(m)` means "stop the chunked sweep at `m`". Both `range_from`
     /// and `range_to` must be `Some` or both `None`; partial sets are
     /// rejected with `FindError::InvalidConfig` at builder time.
-    pub range_to: Option<u64>,
+    ///
+    /// **Type: `u128`.** See [`range_from`](Self::range_from) for rationale.
+    pub range_to: Option<u128>,
     /// Optional Bitcoin mainnet target address (P2PKH 0x00 or P2SH 0x05).
     ///
     /// When `Some(addr)`, the orchestrator switches from variant-keyed
@@ -263,13 +277,16 @@ impl Config {
     /// Sets the explicit scalar range [`from`, `to`]. Both inclusive.
     ///
     /// Both must be provided and `from <= to`. `to` is also implicit by
-    /// the `u64` ceiling; sizes above `u64::MAX` are not representable.
+    /// the `u128` scalar range; values above `u64::MAX = 2^64 - 1`
+    /// are accepted so the address-discovery slice can sweep
+    /// user's `0x4…` style inputs (see ADR-0011).
     ///
     /// When `target_address` is also `Some`, the range scopes the hash40
-    /// sweep; when it's `None`, the range scopes the variant-keyed sweep.
+    /// sweep; when it's `None`, the range scopes the variant-keyed sweep
+    /// (which still uses `u64` internally — see ADR-0011).
     ///
     /// Returns `FindError::InvalidConfig` on validation failure.
-    pub fn try_with_range(mut self, from: u64, to: u64) -> Result<Self> {
+    pub fn try_with_range(mut self, from: u128, to: u128) -> Result<Self> {
         if from > to {
             return Err(FindError::InvalidConfig(format!(
                 "range_from ({from}) must be <= range_to ({to})"
